Add v1.6
"Changeup installation mechanism. Uses cat commands to c/p into service.sh
file instead of directly using scripts in service.d. The plan from here on
is to convert this into a magisk version of an aroma installer, where you use
the volume keys to select the features you want and the installer creates the
script for you"

Signed-off-by: kingbri <bdashore3@gmail.com>